vcubicbezierpath.cpp: fix build with 64-bit long double

+++ src/libs/vgeometry/vcubicbezierpath.cpp
@@ -332,7 +332,7 @@ auto VCubicBezierPath::CountSubSpl(vsizetype size) -> vsizetype
         return 0;
     }
 #if QT_VERSION >= QT_VERSION_CHECK(6, 0, 0)
-    return qFloor(qAbs((size - 4) / 3.0L + 1));
+    return qFloor(qAbs((static_cast<long double>(size) - 4) / 3.0L + 1));
 #else
     return qFloor(qAbs((size - 4) / 3.0 + 1));
 #endif
